+++ packages/rsi-interpretation-player/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.30](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-player@0.3.29...@akkadu/rsi-interpretation-player@0.3.30) (2021-07-29)
+
+**Note:** Version bump only for package @akkadu/rsi-interpretation-player
+
+
+
+
+
+## [0.3.29](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-player@0.3.28...@akkadu/rsi-interpretation-player@0.3.29) (2021-07-27)
+
+**Note:** Version bump only for package @akkadu/rsi-interpretation-player
+
+
+
+
+
 ## [0.3.28](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-interpretation-player@0.3.27...@akkadu/rsi-interpretation-player@0.3.28) (2021-07-21)
 
 **Note:** Version bump only for package @akkadu/rsi-interpretation-player