+++ packages/rsi-vanilla/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.19](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.18...@akkadu/rsi-vanilla@1.0.19) (2021-07-29)
+
+**Note:** Version bump only for package @akkadu/rsi-vanilla
+
+
+
+
+
+## [1.0.18](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.17...@akkadu/rsi-vanilla@1.0.18) (2021-07-27)
+
+**Note:** Version bump only for package @akkadu/rsi-vanilla
+
+
+
+
+
 ## [1.0.17](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-vanilla@1.0.16...@akkadu/rsi-vanilla@1.0.17) (2021-07-21)
 
 **Note:** Version bump only for package @akkadu/rsi-vanilla